Bugfix: add a lock system in order to prevent, in multiprocess mode, two user creations at the same time.
git-svn-id: file:///home/code/svn/ovd/trunk@6190 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ OvdServer/ovd/Role/ApplicationServer/Platform/Linux/User.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2010 Ulteo SAS
+# Copyright (C) 2010-2011 Ulteo SAS
 # http://www.ulteo.com
 # Author Laurent CLOUET <laurent@ulteo.com> 2010
 # Author Julien LANGLOIS <julien@ulteo.com> 2010
+# Author David LECHEVALIER <david@ulteo.com> 2011
 #
 # This program is free software; you can redistribute it and/or 
 # modify it under the terms of the GNU General Public License
@@ -20,6 +21,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import commands
+import fcntl
 import locale
 import pwd
 import xrdp
@@ -28,8 +30,30 @@ from ovd.Logger import Logger
 from ovd.Role.ApplicationServer.User import User as AbstractUser
 
 
+class FileLock():
+	def __init__(self, path):
+		self.path = path
+
+	def acquire(self):
+		try:
+			self.fp = open(self.path, 'w')
+			self.lock = fcntl.flock(self.fp.fileno(),fcntl.LOCK_EX)
+		except Exception, err:
+			Logger.error("Error while acquiring user lock(%s)"%(str(err)))
+
+	def release(self):
+		try:
+			self.lock = fcntl.flock(self.fp.fileno(),fcntl.LOCK_UN)
+			self.fp.close()
+		except Exception, e:
+			Logger.error("Error while releasing user lock(%s)"%(str(err)))
+
+
+
 class User(AbstractUser):
 	def create(self):
+		lock = FileLock("/tmp/user.lock")
+		
 		cmd = u"useradd -m -k /dev/null"
 		if self.infos.has_key("displayName"):
 			cmd+= u" --comment '%s,,,'"%(self.infos["displayName"])
@@ -41,18 +65,55 @@ class User(AbstractUser):
 		
 		cmd+= u" "+self.name
 		
-		s,o = commands.getstatusoutput(cmd.encode(locale.getpreferredencoding()))
-		if s != 0:
-			Logger.error("userAdd return %d (%s)"%(s, o))
-			return False
+		retry = 5
+	        while retry !=0:
+			print "on try"
+		        if retry < 0:
+		                Logger.error("ERROR: unable to add a new user")
+		        lock.acquire()
+		        s,o = commands.getstatusoutput(cmd.encode(locale.getpreferredencoding()))
+		        lock.release()
+		        if s == 0:
+		                break
+			
+		        Logger.debug("Add user :retry %i"%(6-retry))
+		        if s == 2304: # user already exist
+		                Logger.error("User %s already exist"%(self.name))
+		                break;
+		        if s == 256: # an other process is creating a user
+		                Logger.error("An other process is creating a user")
+		                retry -=1
+		                time.sleep(0.2)
+		                continue
+		        if s != 0:
+		                Logger.error("UserAdd return %d (%s)"%(s, o))
+		                return False
+
 		
 		
 		if self.infos.has_key("password"):
 			cmd = 'echo "%s:%s" | chpasswd'%(self.name, self.infos["password"])
-			s,o = commands.getstatusoutput(cmd)
-			if s != 0:
-				Logger.error("chpasswd return %d (%s)"%(s, o))
-				return False
+			retry = 5
+			while retry !=0:
+				if retry < 0:
+					Logger.error("ERROR: unable to add a new user")
+					return False
+				lock.acquire()
+				s,o = commands.getstatusoutput(cmd)
+				lock.release()
+				if s == 0:
+				        break
+				
+			        Logger.debug("Chpasswd of %s:retry %i"%(self.name, 6-retry))
+				if s == 256 or s == 2560: # an other process is creating a user
+				        Logger.debug("An other process is creating a user")
+				        retry -=1
+				        time.sleep(0.2)
+				        continue
+				if s != 0:
+				        Logger.error("chpasswd return %d (%s)"%(s, o))
+				        return False
+
 		
 		return self.post_create()
 	
@@ -86,13 +147,28 @@ class User(AbstractUser):
 	
 	
 	def destroy(self):
+                lock = FileLock("/tmp/user.lock")
+
 		cmd = "userdel --force  --remove %s"%(self.name)
-		
-		s,o = commands.getstatusoutput(cmd)
-		if s == 3072:
-			Logger.debug("mail dir error: '%s' return %d => %s"%(str(cmd), s, o))
-		elif s != 0:
-			Logger.error("userdel return %d (%s)"%(s, o))
-			return False
-		
+		retry = 5
+		while retry !=0:
+			lock.acquire()
+			s,o = commands.getstatusoutput(cmd)
+			lock.release()
+		        if s == 0:
+		                return True
+                        if s == 3072:
+                                Logger.debug("mail dir error: '%s' return %d => %s"%(str(cmd), s, o))
+                                return True
+
+		        Logger.debug("User delete of %s: retry %i"%(self.name, 6-retry))
+		        if s == 256 or s == 2560: # an other process is creating a user
+		                Logger.debug("An other process is creating a user")
+		                retry -=1
+		                time.sleep(0.2)
+		                continue
+		        if s != 0:
+		                Logger.error("userdel return %d (%s)"%(s, o))
+		                return False
+
 		return True
